Program reads and analyzes a log file
This program is an independent project where I use JavaFX and SceneBuilder to visualize data from a log file. The program reads the txt file upon opening and allows the user to select several filters to analyze the data. It's an ongoing project and I hope to add more elements as I learn.